Extracted logic from predict and evaluate endpoints into evaluation handler. Evaluation handler now defines a method that allows to trigger prediction and evaluation at once.
